requirements.txt
whisper
requests
pyttsx3
python-dotenv
